+++ R/show_cn_profile.R
@@ -10,7 +10,11 @@
 #' is a `data.frame`, a column called `sample` must exist.
 #' @param show_n number of samples to show, this is used for checking.
 #' @param show_title if `TRUE`, show title for multiple samples.
+#' @param show_labels one of `NULL`, "s" (for labelling short segments < 1e7)
+#' or "a" (all segments).
 #' @param chrs chromosomes start with 'chr'.
+#' @param position a position range, e.g. `"chr1:3218923-116319008"`. Only data
+#' overlaps with this range will be shown.
 #' @param genome_build genome build version, used when `data` is a `data.frame`, should be 'hg19' or 'hg38'.
 #' @param nrow number of rows in the plot grid when multiple samples are selected.
 #' @param ncol number of columns in the plot grid when multiple samples are selected.
@@ -27,12 +31,21 @@
 #'
 #' p <- show_cn_profile(cn, nrow = 2, ncol = 1)
 #' p
+#' p2 <- show_cn_profile(cn, nrow = 2, ncol = 1,
+#'                       position = "chr1:3218923-116319008")
+#' p2
 #' @testexamples
 #' expect_s3_class(p, "ggplot")
-show_cn_profile <- function(data, samples = NULL, show_n = NULL, show_title = FALSE,
+#' expect_s3_class(p2, "ggplot")
+show_cn_profile <- function(data, samples = NULL,
+                            show_n = NULL, show_title = FALSE,
+                            show_labels = NULL,
                             chrs = paste0("chr", 1:22),
+                            position = NULL,
                             genome_build = c("hg19", "hg38", "mm10"),
-                            nrow = NULL, ncol = NULL, return_plotlist = FALSE) {
+                            ylim = NULL,
+                            nrow = NULL, ncol = NULL,
+                            return_plotlist = FALSE) {
   stopifnot(is.data.frame(data) | inherits(data, "CopyNumber"))
   if (is.data.frame(data)) {
     if (is.null(samples)) {
@@ -66,6 +79,26 @@ show_cn_profile <- function(data, samples = NULL, show_n = NULL, show_title = FA
 
   data <- data[data$chromosome %in% chrs]
 
+  if (!is.null(position)) {
+    stopifnot(length(position) == 1L)
+    position <- unlist(strsplit(position, split = "[:-]"))
+    position_df <- data.table::data.table(
+      chromosome = position[1],
+      start = as.integer(position[2]),
+      end = as.integer(position[3])
+    )
+    data.table::setkey(position_df, chromosome, start, end)
+    data_range <- data.table::foverlaps(data, position_df)[!is.na(start)]
+    if (nrow(data_range) < 1) {
+      message("No data in specified position can be found.")
+      return(invisible(NULL))
+    } else {
+      data <- data_range[, c(-2, -3)]
+      colnames(data)[2:3] <- c("start", "end")
+      chrs <- position[1]
+    }
+  }
+
   # Get plot data
   coord_df <- build_chrom_coordinate(genome_build, chrs) %>%
     dplyr::mutate(labels = sub("chr", "", .data$chrom))
@@ -84,7 +117,10 @@ show_cn_profile <- function(data, samples = NULL, show_n = NULL, show_title = FA
 
   if (!"sample" %in% colnames(merge_df)) {
     # Plot as a single sample
-    gg <- plot_cn_profile(merge_df, coord_df)
+    gg <- plot_cn_profile(merge_df, coord_df, show_labels = show_labels)
+    if (!is.null(ylim)) {
+      gg <- gg + ggplot2::ylim(ylim)
+    }
   } else {
     # Plot mutiple samples
     gg_df <- merge_df %>%
@@ -92,7 +128,8 @@ show_cn_profile <- function(data, samples = NULL, show_n = NULL, show_title = FA
       tidyr::nest() %>%
       dplyr::mutate(gg = purrr::map(.data$data,
         plot_cn_profile,
-        coord_df = coord_df
+        coord_df = coord_df,
+        show_labels = show_labels
       ))
 
     ## group_by does not maintain sample order
@@ -110,6 +147,9 @@ show_cn_profile <- function(data, samples = NULL, show_n = NULL, show_title = FA
     gg_list <- gg_df$gg
     if (show_title) {
       gg_list <- purrr::map2(gg_list, gg_df$sample, ~ .x + labs(title = .y))
+    }
+    if (!is.null(ylim)) {
+      gg_list <- purrr::map2(gg_list, gg_df$sample, ~ .x + ylim(ylim))
     }
     names(gg_list) <- gg_df$sample
     if (return_plotlist) {
@@ -143,13 +183,32 @@ build_chrom_coordinate <- function(genome_build, chrs) {
       dplyr::pull(.data$lab_loc))
 }
 
-plot_cn_profile <- function(plot_df, coord_df) {
-  ggplot() +
+plot_cn_profile <- function(plot_df, coord_df, show_labels = NULL) {
+  p <- ggplot() +
     geom_segment(aes(
       x = .data$start, xend = .data$end,
       y = .data$segVal, yend = .data$segVal,
       color = .data$segType
-    ), data = plot_df) +
+    ), data = plot_df)
+
+  if (!is.null(show_labels)) {
+    p <- p +
+      ggrepel::geom_label_repel(
+        aes(x = (.data$start + .data$end) / 2,
+            y = .data$segVal + 0.05,
+            label = .data$segVal),
+        size = 2,
+        data = if (show_labels == "s") {
+          plot_df %>%
+            dplyr::filter(.data$end - .data$start < 1e7)
+        } else {
+          plot_df
+        },
+        label.padding = unit(0.1, "lines"),
+        point.padding = NA,
+        nudge_y = 0.3)
+  }
+  p +
     geom_vline(aes(xintercept = .data$x_start), linetype = "dotted", data = coord_df) +
     geom_vline(xintercept = coord_df$x_end[nrow(coord_df)], linetype = "dotted") +
     scale_x_continuous(breaks = coord_df$lab_loc, labels = coord_df$labels) +